Dynamnic Time Warping
This project represents the DTW algorithm and how it can be applied in measuring similarity for 2
different series which may differ in length. Via a process of simulation, the relation of similarity measure
that is DTW-based is displayed as with 2 other measures that are well known i.e. the Spearman’s and
the Pearson’s correlation coefficients. Specifically, as demonstrated that takes greater values when the
other 2 measures are great in terms that are absolute. To add to this a dataset which is composed of
NASDAQ closing prices and BTC closing prices and applications of the of aforementioned have been
shown. To start with via a rolling basis, the evolution of has been investigated together with the
volatility and the Pearson’s correlation. Outcomes displayed that in times of high volatility similarities
within the series examined increase. The other thing, a comparison of the average similarities within the
investigated across various classes has been done. Outcomes are different. From this investigation we
were able to find out that the similarity between the two sets of time series data is statistically
significant.